+++ x/bank/keeper/keeper_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/suite"
 
+	"cosmossdk.io/core/header"
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/log"
 	"cosmossdk.io/math"
@@ -31,7 +32,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/query"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	vesting "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
-	"github.com/cosmos/cosmos-sdk/x/bank/exported"
 	"github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	banktestutil "github.com/cosmos/cosmos-sdk/x/bank/testutil"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -1476,7 +1476,7 @@ func (suite *KeeperTestSuite) TestSpendableCoins() {
 	suite.mockSpendableCoins(ctx, acc1)
 	require.Equal(origCoins[0], suite.bankKeeper.SpendableCoin(ctx, accAddrs[1], "stake"))
 
-	ctx = ctx.WithBlockTime(now.Add(12 * time.Hour))
+	ctx = ctx.WithHeaderInfo(header.Info{Time: now.Add(12 * time.Hour)})
 	suite.mockSpendableCoins(ctx, vacc)
 	require.Equal(origCoins.Sub(lockedCoins...), suite.bankKeeper.SpendableCoins(ctx, accAddrs[0]))
 
@@ -1508,7 +1508,7 @@ func (suite *KeeperTestSuite) TestVestingAccountSend() {
 	suite.mockFundAccount(accAddrs[0])
 	require.NoError(banktestutil.FundAccount(ctx, suite.bankKeeper, accAddrs[0], sendCoins))
 	// require that all vested coins are spendable plus any received
-	ctx = ctx.WithBlockTime(now.Add(12 * time.Hour))
+	ctx = ctx.WithHeaderInfo(header.Info{Time: now.Add(12 * time.Hour)})
 	suite.mockSendCoins(ctx, vacc, accAddrs[1])
 	require.NoError(suite.bankKeeper.SendCoins(ctx, accAddrs[0], accAddrs[1], sendCoins))
 	require.Equal(origCoins, suite.bankKeeper.GetAllBalances(ctx, accAddrs[0]))
@@ -1543,7 +1543,7 @@ func (suite *KeeperTestSuite) TestPeriodicVestingAccountSend() {
 	require.NoError(banktestutil.FundAccount(ctx, suite.bankKeeper, accAddrs[0], sendCoins))
 
 	// require that all vested coins are spendable plus any received
-	ctx = ctx.WithBlockTime(now.Add(12 * time.Hour))
+	ctx = ctx.WithHeaderInfo(header.Info{Time: now.Add(12 * time.Hour)})
 	suite.mockSendCoins(ctx, vacc, accAddrs[1])
 	require.NoError(suite.bankKeeper.SendCoins(ctx, accAddrs[0], accAddrs[1], sendCoins))
 	require.Equal(origCoins, suite.bankKeeper.GetAllBalances(ctx, accAddrs[0]))
@@ -1640,7 +1640,7 @@ func (suite *KeeperTestSuite) TestDelegateCoins() {
 	suite.mockFundAccount(accAddrs[1])
 	require.NoError(banktestutil.FundAccount(ctx, suite.bankKeeper, accAddrs[1], origCoins))
 
-	ctx = ctx.WithBlockTime(now.Add(12 * time.Hour))
+	ctx = ctx.WithHeaderInfo(header.Info{Time: now.Add(12 * time.Hour)})
 
 	// require the ability for a non-vesting account to delegate
 	suite.mockDelegateCoins(ctx, acc1, holderAcc)
@@ -1698,7 +1698,7 @@ func (suite *KeeperTestSuite) TestUndelegateCoins() {
 	suite.mockFundAccount(accAddrs[1])
 	require.NoError(banktestutil.FundAccount(ctx, suite.bankKeeper, accAddrs[1], origCoins))
 
-	ctx = ctx.WithBlockTime(now.Add(12 * time.Hour))
+	ctx = ctx.WithHeaderInfo(header.Info{Time: now.Add(12 * time.Hour)})
 
 	// require the ability for a non-vesting account to delegate
 	suite.mockDelegateCoins(ctx, acc1, holderAcc)
@@ -2340,66 +2340,6 @@ func (suite *KeeperTestSuite) TestGetAllSendEnabledEntries() {
 		actual := bankKeeper.GetAllSendEnabledEntries(ctx)
 		require.Len(actual, 0)
 	})
-}
-
-type mockSubspace struct {
-	ps banktypes.Params
-}
-
-func (ms mockSubspace) GetParamSet(ctx sdk.Context, ps exported.ParamSet) {
-	*ps.(*banktypes.Params) = ms.ps
-}
-
-func (suite *KeeperTestSuite) TestMigrator_Migrate3to4() {
-	bankKeeper := suite.bankKeeper
-	ctx := sdk.UnwrapSDKContext(suite.ctx)
-	require := suite.Require()
-
-	for _, def := range []bool{true, false} {
-		params := banktypes.Params{DefaultSendEnabled: def}
-		require.NoError(bankKeeper.SetParams(ctx, params))
-
-		suite.T().Run(fmt.Sprintf("default %t does not change", def), func(t *testing.T) {
-			legacySubspace := func(ps banktypes.Params) mockSubspace {
-				return mockSubspace{ps: ps}
-			}(banktypes.NewParams(def))
-
-			migrator := keeper.NewMigrator(bankKeeper, legacySubspace)
-			require.NoError(migrator.Migrate3to4(ctx))
-
-			actual := bankKeeper.GetParams(ctx)
-			require.Equal(params.DefaultSendEnabled, actual.DefaultSendEnabled)
-		})
-	}
-
-	for _, def := range []bool{true, false} {
-		params := banktypes.Params{
-			SendEnabled: []*banktypes.SendEnabled{
-				{Denom: fmt.Sprintf("truecoin%t", def), Enabled: true},
-				{Denom: fmt.Sprintf("falsecoin%t", def), Enabled: false},
-			},
-		}
-
-		require.NoError(bankKeeper.SetParams(ctx, params))
-
-		suite.T().Run(fmt.Sprintf("default %t send enabled info moved to store", def), func(t *testing.T) {
-			legacySubspace := func(ps banktypes.Params) mockSubspace {
-				return mockSubspace{ps: ps}
-			}(banktypes.NewParams(def))
-
-			migrator := keeper.NewMigrator(bankKeeper, legacySubspace)
-			require.NoError(migrator.Migrate3to4(ctx))
-
-			newParams := bankKeeper.GetParams(ctx)
-			require.Len(newParams.SendEnabled, 0) //nolint:staticcheck // we're testing the old way here
-			require.Equal(def, newParams.DefaultSendEnabled)
-
-			for _, se := range params.SendEnabled {
-				actual := bankKeeper.IsSendEnabledDenom(ctx, se.Denom)
-				require.Equal(se.Enabled, actual, se.Denom)
-			}
-		})
-	}
 }
 
 func (suite *KeeperTestSuite) TestSetParams() {